add: script for converting animated webp to webm
this script is able to convert an animated webp to
webm.
This aims to make Whatsapp animated stickers usable
on Telegram, so there are some constraints I applied:
- webm should be 3 s long (maximum)
- webm size should be 512x512 px
- webm should be 256 Kb large (maximum)

This is not able to keep the transparent background
from some animated webp, but I'll be looking into
it at some point.